refactor: move repetitive logic to a separate class
Now everything that is common between two VC (AddGoalViewController и GoalDetailsViewController) -> GoalViewController
